requirements.txt
# --- Core ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1

# --- Data ---
pandas>=2.2.0
yfinance>=0.2.40
numpy>=1.26.0

# --- AI Model ---
# Chronos-2 requires >=2.0. torch is installed separately in Dockerfile for CPU build.
chronos-forecasting>=2.0

# --- Visualization ---
matplotlib>=3.9.0

# --- Telegram (stdlib http.client is used, no extra lib needed) ---
